requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
alembic==1.8.1
anyio==3.6.1
asgi-lifespan==1.0.1
async-timeout==4.0.2
asyncpg==0.26.0
attrs==22.1.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.1.0
commonmark==0.9.1
coolname==1.1.0
croniter==1.3.5
cryptography==36.0.2
docker==6.0.0
fastapi==0.81.0
frozenlist==1.3.1
fsspec==2022.8.0
google-auth==2.11.0
greenlet==1.1.3
griffe==0.21.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
jsonpatch==1.32
jsonpointer==2.3
kubernetes==24.2.0
Mako==1.2.2
MarkupSafe==2.1.1
multidict==6.0.2
oauthlib==3.2.0
orjson==3.8.0
packaging==21.3
pathspec==0.9.0
pendulum==2.1.2
prefect==2.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.0
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-slugify==6.1.2
pytz==2022.2.1
pytzdata==2020.1
PyYAML==6.0
readchar==4.0.2
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==12.5.1
rsa==4.9
six==1.16.0
slack-sdk==3.18.1
sniffio==1.2.0
snowflake==0.0.3
SQLAlchemy==1.4.40
starlette==0.19.1
text-unidecode==1.3
toml==0.10.2
typer==0.6.1
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
websocket-client==1.4.0
yarl==1.8.1
zipp==3.8.1

# https://github.com/snowflakedb/snowflake-connector-python/blob/main/tested_requirements/requirements_38.reqs
# Generated on: Python 3.8.13
asn1crypto==1.5.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==36.0.2
filelock==3.8.0
idna==3.3
oscrypto==1.3.0
pycparser==2.21
pycryptodomex==3.15.0
PyJWT==2.4.0
pyOpenSSL==22.0.0
pytz==2022.2.1
requests==2.28.1
typing_extensions==4.3.0
urllib3==1.26.12
snowflake-connector-python==2.7.12
